services/review-service/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
python-dotenv
requests      
gunicorn          
psycopg2-binary
pytz
python-dateutil
